# public/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"error": {},
					"data": {},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"error",
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "employee-review",
		"version": "1.0.0",
		"description": "Employee Review apis",
		"license": {
			"name": "license"
		},
		"contact": {
			"name": "Sandeep Ghildiyal"
		}
	},
	"paths": {
		"/api/employee/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Login manager",
				"tags": [
					"Employees"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/employee/addEmployee": {
			"post": {
				"operationId": "AddEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Add a  employee",
				"tags": [
					"Employees"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"joiningDate": {
										"type": "string"
									},
									"employeeId": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"employeeName": {
										"type": "string"
									}
								},
								"required": [
									"joiningDate",
									"employeeId",
									"email",
									"employeeName"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/employee/update": {
			"post": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "update a  employee",
				"tags": [
					"Employees"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"joiningDate": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"employeeName": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"joiningDate",
									"email",
									"employeeName",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/employee/getEmployebyId": {
			"get": {
				"operationId": "GetEmployebyId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "update a  employee",
				"tags": [
					"Employees"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/employee/getAllEmployee": {
			"get": {
				"operationId": "GetAllEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Get all user, require admin token",
				"tags": [
					"Employees"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/employee/delete": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Employees"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/employee/generateFeedback": {
			"post": {
				"operationId": "GenerateFeedback",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Employees"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"metrics": {
										"properties": {
											"problemSolving": {
												"type": "number",
												"format": "double"
											},
											"communication": {
												"type": "number",
												"format": "double"
											},
											"punctuality": {
												"type": "number",
												"format": "double"
											},
											"teamwork": {
												"type": "number",
												"format": "double"
											},
											"productivity": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"problemSolving",
											"communication",
											"punctuality",
											"teamwork",
											"productivity"
										],
										"type": "object"
									},
									"periodTo": {
										"type": "string"
									},
									"periodFrom": {
										"type": "string"
									},
									"employeeId": {
										"type": "string"
									}
								},
								"required": [
									"metrics",
									"periodTo",
									"periodFrom",
									"employeeId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/employee/compareEmployeeFeedback": {
			"get": {
				"operationId": "CompareEmployeeFeedback",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Compare two employees' feedback",
				"tags": [
					"Employees"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "employee1Id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "employee2Id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}